+++ competencias-JPA/src/main/java/es/aragon/iacs/competencias/dao/ICompResultadosDAO.java
@@ -5,6 +5,6 @@ import es.aragon.iacs.competencias.jpa.CompResultados;
 
 public interface ICompResultadosDAO {
 
-	void sumarValoracion(Integer idEvaluacion, String dniEvaluado, Integer grupoev,String codcomp, Integer idnivel, Integer valoracion);
+	void sumarValoracion(Integer idEvaluacion, String dniEvaluado, Integer grupoev,String codcomp, Integer idnivel, float valoracion);
 	List<CompResultados> evaluacionCalculada(Integer idEvaluacion);
 }
